pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adcs-lab"
version = "0.1.0"
description = "Active Directory Certificate Services Attack & Defence Lab (simulated)"
authors = [{name = "Lab Maintainers", email = "maintainers@example.local"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.7",
]
license = {text = "MIT"}

[project.scripts]
adcs-lab = "adcs_lab.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flake8>=6.0",
    "black>=24.3.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12.8",
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_unused_configs = true
exclude = ["tests/.*"]

[tool.setuptools.packages.find]
where = ["src"]
